combine profile and time_trace functions in ensemble DA diagnostics script.
